backend/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
sentence-transformers
faiss-cpu
mistralai
python-dotenv
requests
joblib
pdfplumber
bcrypt
python-multipart
